+++ NEWS.md
@@ -7,3 +7,5 @@
 
 - Lazy covariance and empirical structure function object have (read-only)
   public properties to favor using the `A.f` syntax.
+
+- Empirical structure function object is an abstract array (with offsets).